+++ src/NotificationThrottle.ts
@@ -1,0 +1,11 @@
+export class NotificationThrottle {
+  #hasNotified = false;
+
+  shouldNotify() {
+    if (this.#hasNotified) {
+      return false;
+    }
+    this.#hasNotified = true;
+    return true;
+  }
+}